ai-llm-validator/requirements.txt
# ──────────────────────────────────────────────
# AI LLM Validator — Python Dependencies
# ──────────────────────────────────────────────
# Install with:  pip install -r requirements.txt
# ──────────────────────────────────────────────

# --- Core ML / NLP ---
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
spacy>=3.5.0

# --- Data / Scientific ---
numpy>=1.24.0
scikit-learn>=1.2.0

# --- Web UI ---
streamlit>=1.24.0

# --- Utilities ---
python-dotenv>=1.0.0
